BankTools_AI/backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
flask-mail==0.9.1
flask-cors==4.0.0
werkzeug==2.3.7
wtforms==3.1.1
email-validator==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.1
gunicorn==21.2.0
pytest==7.4.3 